[Decomp][Spell] Cohesion split: extract SpellEffectSkillEnchantPet handlers to SpellEffectSkillEnchantPet.cpp
