+++ src/game/WorldHandlers/SpellEffectSkillEnchantPet.cpp
@@ -1,0 +1,1215 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+
+
+#include "Common.h"
+#include "Database/DatabaseEnv.h"
+#include "WorldPacket.h"
+#include "Opcodes.h"
+#include "Log.h"
+#include "UpdateMask.h"
+#include "World.h"
+#include "ObjectMgr.h"
+#include "SpellMgr.h"
+#include "Player.h"
+#include "SkillExtraItems.h"
+#include "Unit.h"
+#include "Spell.h"
+#include "DynamicObject.h"
+#include "SpellAuras.h"
+#include "Group.h"
+#include "UpdateData.h"
+#include "MapManager.h"
+#include "ObjectAccessor.h"
+#include "SharedDefines.h"
+#include "Pet.h"
+#include "GameObject.h"
+#include "GossipDef.h"
+#include "Creature.h"
+#include "Totem.h"
+#include "CreatureAI.h"
+#include "BattleGround/BattleGroundMgr.h"
+#include "BattleGround/BattleGround.h"
+#include "BattleGround/BattleGroundEY.h"
+#include "BattleGround/BattleGroundWS.h"
+#include "Language.h"
+#include "SocialMgr.h"
+#include "VMapFactory.h"
+#include "Util.h"
+#include "TemporarySummon.h"
+#include "ScriptMgr.h"
+#include "SkillDiscovery.h"
+#include "Formulas.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "CellImpl.h"
+#include "G3D/Vector3.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+/**
+ * @brief Teaches a spell to the target player or pet.
+ *
+ * @param eff_idx The effect index containing the learned spell id.
+ */
+void Spell::EffectLearnSpell(SpellEffectIndex eff_idx)
+{
+    if (!unitTarget)
+    {
+        return;
+    }
+
+    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
+    {
+        if (m_caster->GetTypeId() == TYPEID_PLAYER)
+        {
+            EffectLearnPetSpell(eff_idx);
+        }
+
+        return;
+    }
+
+    Player* player = (Player*)unitTarget;
+    uint32 spellToLearn = (m_spellInfo->Id == SPELL_ID_GENERIC_LEARN) ? damage : m_spellInfo->EffectTriggerSpell[eff_idx];
+
+    // special case for paladin SoR 20154 (non-judgement version)
+    if (m_spellInfo->Id == 10321)
+    {
+        player->removeSpell(20154, true, false);
+    }
+
+    player->learnSpell(spellToLearn, false);
+
+    if (WorldObject const* caster = GetCastingObject())
+    {
+        DEBUG_LOG("Spell: %s has learned spell %u from %s", player->GetGuidStr().c_str(), spellToLearn, caster->GetGuidStr().c_str());
+    }
+}
+
+/**
+ * @brief Dispels matching auras from the target and sends result logs.
+ *
+ * @param eff_idx The dispel effect index.
+ */
+void Spell::EffectDispel(SpellEffectIndex eff_idx)
+{
+    if (!unitTarget)
+    {
+        return;
+    }
+
+    // Fill possible dispel list
+    std::list <std::pair<SpellAuraHolder* , uint32> > dispel_list;
+
+    // Create dispel mask by dispel type
+    uint32 dispel_type = m_spellInfo->EffectMiscValue[eff_idx];
+    uint32 dispelMask  = GetDispellMask(DispelType(dispel_type));
+    Unit::SpellAuraHolderMap const& auras = unitTarget->GetSpellAuraHolderMap();
+    for (Unit::SpellAuraHolderMap::const_iterator itr = auras.begin(); itr != auras.end(); ++itr)
+    {
+        SpellAuraHolder* holder = itr->second;
+        if ((1 << holder->GetSpellProto()->Dispel) & dispelMask)
+        {
+            if (holder->GetSpellProto()->Dispel == DISPEL_MAGIC)
+            {
+                bool positive = true;
+                if (!holder->IsPositive())
+                {
+                    positive = false;
+                }
+                else
+                {
+                    positive = (holder->GetSpellProto()->AttributesEx & SPELL_ATTR_EX_CANT_BE_REFLECTED) == 0;
+                }
+
+                // do not remove positive auras if friendly target
+                //               negative auras if non-friendly target
+                if (positive == unitTarget->IsFriendlyTo(m_caster))
+                {
+                    continue;
+                }
+            }
+            dispel_list.push_back(std::pair<SpellAuraHolder* , uint32>(holder, holder->GetStackAmount()));
+        }
+    }
+    // Ok if exist some buffs for dispel try dispel it
+    if (!dispel_list.empty())
+    {
+        std::list<std::pair<SpellAuraHolder* , uint32> > success_list; // (spell_id,casterGuid)
+        std::list < uint32 > fail_list;                     // spell_id
+
+        // some spells have effect value = 0 and all from its by meaning expect 1
+        if (!damage)
+        {
+            damage = 1;
+        }
+
+        // Dispel N = damage buffs (or while exist buffs for dispel)
+        for (int32 count = 0; count < damage && !dispel_list.empty(); ++count)
+        {
+            // Random select buff for dispel
+            std::list<std::pair<SpellAuraHolder* , uint32> >::iterator dispel_itr = dispel_list.begin();
+            std::advance(dispel_itr, urand(0, dispel_list.size() - 1));
+
+            SpellAuraHolder* holder = dispel_itr->first;
+
+            dispel_itr->second -= 1;
+
+            // remove entry from dispel_list if nothing left in stack
+            if (dispel_itr->second == 0)
+            {
+                dispel_list.erase(dispel_itr);
+            }
+
+            SpellEntry const* spellInfo = holder->GetSpellProto();
+            // Base dispel chance
+            // TODO: possible chance depend from spell level??
+            int32 miss_chance = 0;
+            // Apply dispel mod from aura caster
+            if (Unit* caster = holder->GetCaster())
+            {
+                if (Player* modOwner = caster->GetSpellModOwner())
+                {
+                    modOwner->ApplySpellMod(spellInfo->Id, SPELLMOD_RESIST_DISPEL_CHANCE, miss_chance, this);
+                }
+            }
+            // Try dispel
+            if (roll_chance_i(miss_chance))
+            {
+                fail_list.push_back(spellInfo->Id);
+            }
+            else
+            {
+                bool foundDispelled = false;
+                for (std::list<std::pair<SpellAuraHolder* , uint32> >::iterator success_iter = success_list.begin(); success_iter != success_list.end(); ++success_iter)
+                {
+                    if (success_iter->first->GetId() == holder->GetId() && success_iter->first->GetCasterGuid() == holder->GetCasterGuid())
+                    {
+                        success_iter->second += 1;
+                        foundDispelled = true;
+                        break;
+                    }
+                }
+                if (!foundDispelled)
+                {
+                    success_list.push_back(std::pair<SpellAuraHolder* , uint32>(holder, 1));
+                }
+            }
+        }
+        // Send success log and really remove auras
+        if (!success_list.empty())
+        {
+            int32 count = success_list.size();
+            WorldPacket data(SMSG_SPELLDISPELLOG, 8 + 8 + 4 + 1 + 4 + count * 5);
+            data << unitTarget->GetPackGUID();              // Victim GUID
+            data << m_caster->GetPackGUID();                // Caster GUID
+            data << uint32(m_spellInfo->Id);                // Dispel spell id
+            data << uint8(0);                               // not used
+            data << uint32(count);                          // count
+            for (std::list<std::pair<SpellAuraHolder* , uint32> >::iterator j = success_list.begin(); j != success_list.end(); ++j)
+            {
+                SpellAuraHolder* dispelledHolder = j->first;
+                data << uint32(dispelledHolder->GetId());   // Spell Id
+                data << uint8(0);                           // 0 - dispelled !=0 cleansed
+                unitTarget->RemoveAuraHolderDueToSpellByDispel(dispelledHolder->GetId(), j->second, dispelledHolder->GetCasterGuid(), m_caster);
+            }
+            m_caster->SendMessageToSet(&data, true);
+
+            // On success dispel
+            // Devour Magic
+            if (m_spellInfo->SpellFamilyName == SPELLFAMILY_WARLOCK && m_spellInfo->Category == SPELLCATEGORY_DEVOUR_MAGIC)
+            {
+                uint32 heal_spell = 0;
+                switch (m_spellInfo->Id)
+                {
+                    case 19505: heal_spell = 19658; break;
+                    case 19731: heal_spell = 19732; break;
+                    case 19734: heal_spell = 19733; break;
+                    case 19736: heal_spell = 19735; break;
+                    case 27276: heal_spell = 27278; break;
+                    case 27277: heal_spell = 27279; break;
+                    default:
+                        DEBUG_LOG("Spell for Devour Magic %d not handled in Spell::EffectDispel", m_spellInfo->Id);
+                        break;
+                }
+                if (heal_spell)
+                {
+                    m_caster->CastSpell(m_caster, heal_spell, true);
+                }
+            }
+        }
+        // Send fail log to client
+        if (!fail_list.empty())
+        {
+            // Failed to dispel
+            WorldPacket data(SMSG_DISPEL_FAILED, 8 + 8 + 4 + 4 * fail_list.size());
+            data << m_caster->GetObjectGuid();              // Caster GUID
+            data << unitTarget->GetObjectGuid();            // Victim GUID
+            data << uint32(m_spellInfo->Id);                // Dispel spell id
+            for (std::list< uint32 >::iterator j = fail_list.begin(); j != fail_list.end(); ++j)
+            {
+                data << uint32(*j);                          // Spell Id
+            }
+            m_caster->SendMessageToSet(&data, true);
+        }
+    }
+}
+
+/**
+ * @brief Enables dual wielding for a player target.
+ */
+void Spell::EffectDualWield(SpellEffectIndex /*eff_idx*/)
+{
+    if (unitTarget && unitTarget->GetTypeId() == TYPEID_PLAYER)
+    {
+        ((Player*)unitTarget)->SetCanDualWield(true);
+    }
+}
+
+/**
+ * @brief Placeholder for pull-style spell effects.
+ */
+void Spell::EffectPull(SpellEffectIndex /*eff_idx*/)
+{
+    // TODO: create a proper pull towards distract spell center for distract
+    DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "WORLD: Spell Effect DUMMY");
+}
+
+/**
+ * @brief Turns and distracts a non-combat unit target.
+ */
+void Spell::EffectDistract(SpellEffectIndex /*eff_idx*/)
+{
+    // Check for possible target
+    if (!unitTarget || unitTarget->IsInCombat())
+    {
+        return;
+    }
+
+    // target must be OK to do this
+    if (unitTarget->hasUnitState(UNIT_STAT_CAN_NOT_REACT))
+    {
+        return;
+    }
+
+    unitTarget->SetFacingTo(unitTarget->GetAngle(m_targets.m_destX, m_targets.m_destY));
+    unitTarget->clearUnitState(UNIT_STAT_MOVING);
+
+    if (unitTarget->GetTypeId() == TYPEID_UNIT)
+    {
+        unitTarget->GetMotionMaster()->MoveDistract(damage * IN_MILLISECONDS);
+    }
+}
+
+/**
+ * @brief Attempts to pickpocket a valid creature target.
+ */
+void Spell::EffectPickPocket(SpellEffectIndex /*eff_idx*/)
+{
+    if (m_caster->GetTypeId() != TYPEID_PLAYER)
+    {
+        return;
+    }
+
+    // victim must be creature and attackable
+    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_UNIT || m_caster->IsFriendlyTo(unitTarget))
+    {
+        return;
+    }
+
+    // victim have to be alive and humanoid or undead
+    if (unitTarget->IsAlive() && (unitTarget->GetCreatureTypeMask() & CREATURE_TYPEMASK_HUMANOID_OR_UNDEAD) != 0)
+    {
+        int32 chance = 10 + int32(m_caster->getLevel()) - int32(unitTarget->getLevel());
+
+        if (chance > irand(0, 19))
+        {
+            // Stealing successful
+            // DEBUG_LOG("Sending loot from pickpocket");
+            ((Player*)m_caster)->SendLoot(unitTarget->GetObjectGuid(), LOOT_PICKPOCKETING);
+        }
+        else
+        {
+            // Reveal action + get attack
+            m_caster->RemoveSpellsCausingAura(SPELL_AURA_MOD_STEALTH);
+            unitTarget->AttackedBy(m_caster);
+        }
+    }
+}
+
+/**
+ * @brief Creates a farsight focus object and switches the player's camera to it.
+ *
+ * @param eff_idx The farsight effect index.
+ */
+void Spell::EffectAddFarsight(SpellEffectIndex eff_idx)
+{
+    if (m_caster->GetTypeId() != TYPEID_PLAYER)
+    {
+        return;
+    }
+
+    int32 duration = GetSpellDuration(m_spellInfo);
+    DynamicObject* dynObj = new DynamicObject;
+
+    // set radius to 0: spell not expected to work as persistent aura
+    if (!dynObj->Create(m_caster->GetMap()->GenerateLocalLowGuid(HIGHGUID_DYNAMICOBJECT), m_caster,
+                        m_spellInfo->Id, eff_idx, m_targets.m_destX, m_targets.m_destY, m_targets.m_destZ, duration, 0, DYNAMIC_OBJECT_FARSIGHT_FOCUS))
+    {
+        delete dynObj;
+        return;
+    }
+
+    m_caster->AddDynObject(dynObj);
+    m_caster->GetMap()->Add(dynObj);
+
+    ((Player*)m_caster)->GetCamera().SetView(dynObj);
+}
+
+/**
+ * @brief Teleports the target near the caster and turns it to face away from the caster.
+ *
+ * @param eff_idx The teleport effect index.
+ */
+void Spell::EffectTeleUnitsFaceCaster(SpellEffectIndex eff_idx)
+{
+    if (!unitTarget)
+    {
+        return;
+    }
+
+    if (unitTarget->IsTaxiFlying())
+    {
+        return;
+    }
+
+    float fx, fy, fz;
+    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
+    {
+        m_targets.getDestination(fx, fy, fz);
+    }
+    else
+    {
+        float dis = GetSpellRadius(sSpellRadiusStore.LookupEntry(m_spellInfo->EffectRadiusIndex[eff_idx]));
+        m_caster->GetClosePoint(fx, fy, fz, unitTarget->GetObjectBoundingRadius(), dis);
+    }
+
+    unitTarget->NearTeleportTo(fx, fy, fz, m_caster->GetOrientation() + M_PI_F, unitTarget == m_caster);
+}
+
+/**
+ * @brief Teaches or updates a skill for a player target.
+ *
+ * @param eff_idx The effect index containing the skill identifier.
+ */
+void Spell::EffectLearnSkill(SpellEffectIndex eff_idx)
+{
+    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
+    {
+        return;
+    }
+
+    if (damage < 0)
+    {
+        return;
+    }
+
+    uint32 skillid =  m_spellInfo->EffectMiscValue[eff_idx];
+    uint16 skillval = ((Player*)unitTarget)->GetPureSkillValue(skillid);
+    ((Player*)unitTarget)->SetSkill(skillid, skillval ? skillval : 1, damage * 75, damage);
+
+    if (WorldObject const* caster = GetCastingObject())
+    {
+        DEBUG_LOG("Spell: %s has learned skill %u (to maxlevel %u) from %s", unitTarget->GetGuidStr().c_str(), skillid, damage * 75, caster->GetGuidStr().c_str());
+    }
+}
+
+/**
+ * @brief Placeholder handler for trade-skill effects.
+ */
+void Spell::EffectTradeSkill(SpellEffectIndex /*eff_idx*/)
+{
+    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
+    {
+        return;
+    }
+    // uint32 skillid =  m_spellInfo->EffectMiscValue[i];
+    // uint16 skillmax = ((Player*)unitTarget)->(skillid);
+    // ((Player*)unitTarget)->SetSkill(skillid,skillval?skillval:1,skillmax+75);
+}
+
+/**
+ * @brief Applies a permanent enchantment to the target item.
+ *
+ * @param eff_idx The effect index providing the enchantment id.
+ */
+void Spell::EffectEnchantItemPerm(SpellEffectIndex eff_idx)
+{
+    if (m_caster->GetTypeId() != TYPEID_PLAYER)
+    {
+        return;
+    }
+    if (!itemTarget)
+    {
+        return;
+    }
+
+    Player* p_caster = (Player*)m_caster;
+
+    // not grow at item use at item case
+    p_caster->UpdateCraftSkill(m_spellInfo->Id);
+
+    uint32 enchant_id = m_spellInfo->EffectMiscValue[eff_idx];
+    if (!enchant_id)
+    {
+        return;
+    }
+
+    SpellItemEnchantmentEntry const* pEnchant = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
+    if (!pEnchant)
+    {
+        return;
+    }
+
+    // item can be in trade slot and have owner diff. from caster
+    Player* item_owner = itemTarget->GetOwner();
+    if (!item_owner)
+    {
+        return;
+    }
+
+    if (item_owner != p_caster && p_caster->GetSession()->GetSecurity() > SEC_PLAYER && sWorld.getConfig(CONFIG_BOOL_GM_LOG_TRADE))
+    {
+        sLog.outCommand(p_caster->GetSession()->GetAccountId(), "GM %s (Account: %u) enchanting(perm): %s (Entry: %d) for player: %s (Account: %u)",
+                        p_caster->GetName(), p_caster->GetSession()->GetAccountId(),
+                        itemTarget->GetProto()->Name1, itemTarget->GetEntry(),
+                        item_owner->GetName(), item_owner->GetSession()->GetAccountId());
+    }
+
+    // remove old enchanting before applying new if equipped
+    item_owner->ApplyEnchantment(itemTarget, PERM_ENCHANTMENT_SLOT, false);
+
+    itemTarget->SetEnchantment(PERM_ENCHANTMENT_SLOT, enchant_id, 0, 0);
+
+    // add new enchanting if equipped
+    item_owner->ApplyEnchantment(itemTarget, PERM_ENCHANTMENT_SLOT, true);
+}
+
+/**
+ * @brief Applies a temporary enchantment to the target item.
+ *
+ * @param eff_idx The effect index providing the enchantment id.
+ */
+void Spell::EffectEnchantItemTmp(SpellEffectIndex eff_idx)
+{
+    if (m_caster->GetTypeId() != TYPEID_PLAYER)
+    {
+        return;
+    }
+    if (!itemTarget)
+    {
+        return;
+    }
+
+    Player* p_caster = (Player*)m_caster;
+
+    uint32 enchant_id = m_spellInfo->EffectMiscValue[eff_idx];
+
+    // Shaman Rockbiter Weapon
+    if (eff_idx == EFFECT_INDEX_0 && m_spellInfo->Effect[EFFECT_INDEX_1] == SPELL_EFFECT_DUMMY)
+    {
+        int32 enchanting_damage = m_currentBasePoints[EFFECT_INDEX_1];
+
+        // enchanting id selected by calculated damage-per-sec stored in Effect[1] base value
+        // with already applied percent bonus from Elemental Weapons talent
+        // Note: damage calculated (correctly) with rounding int32(float(v)) but
+        // RW enchantments applied damage int32(float(v)+0.5), this create  0..1 difference sometime
+        switch (enchanting_damage)
+        {
+                // Rank 1
+            case  2: enchant_id =   29; break;              //  0% [ 7% ==  2, 14% == 2, 20% == 2]
+                // Rank 2
+            case  4: enchant_id =    6; break;              //  0% [ 7% ==  4, 14% == 4]
+            case  5: enchant_id = 3025; break;              // 20%
+                // Rank 3
+            case  6: enchant_id =    1; break;              //  0% [ 7% ==  6, 14% == 6]
+            case  7: enchant_id = 3027; break;              // 20%
+                // Rank 4
+            case  9: enchant_id = 3032; break;              //  0% [ 7% ==  6]
+            case 10: enchant_id =  503; break;              // 14%
+            case 11: enchant_id = 3031; break;              // 20%
+                // Rank 5
+            case 15: enchant_id = 3035; break;              // 0%
+            case 16: enchant_id = 1663; break;              // 7%
+            case 17: enchant_id = 3033; break;              // 14%
+            case 18: enchant_id = 3034; break;              // 20%
+                // Rank 6
+            case 28: enchant_id = 3038; break;              // 0%
+            case 29: enchant_id =  683; break;              // 7%
+            case 31: enchant_id = 3036; break;              // 14%
+            case 33: enchant_id = 3037; break;              // 20%
+                // Rank 7
+            case 40: enchant_id = 3041; break;              // 0%
+            case 42: enchant_id = 1664; break;              // 7%
+            case 45: enchant_id = 3039; break;              // 14%
+            case 48: enchant_id = 3040; break;              // 20%
+                // Rank 8
+            case 49: enchant_id = 3044; break;              // 0%
+            case 52: enchant_id = 2632; break;              // 7%
+            case 55: enchant_id = 3042; break;              // 14%
+            case 58: enchant_id = 3043; break;              // 20%
+                // Rank 9
+            case 62: enchant_id = 2633; break;              // 0%
+            case 66: enchant_id = 3018; break;              // 7%
+            case 70: enchant_id = 3019; break;              // 14%
+            case 74: enchant_id = 3020; break;              // 20%
+            default:
+                sLog.outError("Spell::EffectEnchantItemTmp: Damage %u not handled in S'RW", enchanting_damage);
+                return;
+        }
+    }
+
+    if (!enchant_id)
+    {
+        sLog.outError("Spell %u Effect %u (SPELL_EFFECT_ENCHANT_ITEM_TEMPORARY) have 0 as enchanting id", m_spellInfo->Id, eff_idx);
+        return;
+    }
+
+    SpellItemEnchantmentEntry const* pEnchant = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
+    if (!pEnchant)
+    {
+        sLog.outError("Spell %u Effect %u (SPELL_EFFECT_ENCHANT_ITEM_TEMPORARY) have nonexistent enchanting id %u ", m_spellInfo->Id, eff_idx, enchant_id);
+        return;
+    }
+
+    // select enchantment duration
+    uint32 duration;
+
+    // rogue family enchantments exception by duration
+    if (m_spellInfo->Id == 38615)
+    {
+        duration = 1800;                                    // 30 mins
+    }
+    // other rogue family enchantments always 1 hour (some have spell damage=0, but some have wrong data in EffBasePoints)
+    else if (m_spellInfo->SpellFamilyName == SPELLFAMILY_ROGUE)
+    {
+        duration = 3600;                                    // 1 hour
+    }
+    // shaman family enchantments
+    else if (m_spellInfo->SpellFamilyName == SPELLFAMILY_SHAMAN)
+    {
+        duration = 1800;                                    // 30 mins
+    }
+    // other cases with this SpellVisual already selected
+    else if (m_spellInfo->SpellVisual == 215)
+    {
+        duration = 1800;                                    // 30 mins
+    }
+    // some fishing pole bonuses
+    else if (m_spellInfo->SpellVisual == 563)
+    {
+        duration = 600;                                      // 10 mins
+    }
+    // shaman rockbiter enchantments
+    else if (m_spellInfo->SpellVisual == 0)
+    {
+        duration = 1800;                                    // 30 mins
+    }
+    else if (m_spellInfo->Id == 29702)
+    {
+        duration = 300;                                     // 5 mins
+    }
+    else if (m_spellInfo->Id == 37360)
+    {
+        duration = 300;                                     // 5 mins
+    }
+    // default case
+    else
+    {
+        duration = 3600;                                    // 1 hour
+    }
+
+    // item can be in trade slot and have owner diff. from caster
+    Player* item_owner = itemTarget->GetOwner();
+    if (!item_owner)
+    {
+        return;
+    }
+
+    if (item_owner != p_caster && p_caster->GetSession()->GetSecurity() > SEC_PLAYER && sWorld.getConfig(CONFIG_BOOL_GM_LOG_TRADE))
+    {
+        sLog.outCommand(p_caster->GetSession()->GetAccountId(), "GM %s (Account: %u) enchanting(temp): %s (Entry: %d) for player: %s (Account: %u)",
+                        p_caster->GetName(), p_caster->GetSession()->GetAccountId(),
+                        itemTarget->GetProto()->Name1, itemTarget->GetEntry(),
+                        item_owner->GetName(), item_owner->GetSession()->GetAccountId());
+    }
+
+    // remove old enchanting before applying new if equipped
+    item_owner->ApplyEnchantment(itemTarget, TEMP_ENCHANTMENT_SLOT, false);
+
+    itemTarget->SetEnchantment(TEMP_ENCHANTMENT_SLOT, enchant_id, duration * 1000, 0);
+
+    // add new enchanting if equipped
+    item_owner->ApplyEnchantment(itemTarget, TEMP_ENCHANTMENT_SLOT, true);
+}
+
+/**
+ * @brief Converts the creature target into a hunter pet for the caster.
+ */
+void Spell::EffectTameCreature(SpellEffectIndex /*eff_idx*/)
+{
+    // Caster must be player, checked in Spell::CheckCast
+    // Spell can be triggered, we need to check original caster prior to caster
+    Player* plr = (Player*)GetAffectiveCaster();
+
+    Creature* creatureTarget = (Creature*)unitTarget;
+
+    // cast finish successfully
+    // SendChannelUpdate(0);
+    finish();
+
+    Pet* pet = new Pet(HUNTER_PET);
+
+    if (!pet->CreateBaseAtCreature(creatureTarget))
+    {
+        delete pet;
+        return;
+    }
+
+    pet->SetOwnerGuid(plr->GetObjectGuid());
+    pet->SetCreatorGuid(plr->GetObjectGuid());
+    pet->setFaction(plr->getFaction());
+    pet->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
+
+    if (plr->IsPvP())
+    {
+        pet->SetPvP(true);
+    }
+
+    if (!pet->InitStatsForLevel(creatureTarget->getLevel()))
+    {
+        sLog.outError("Pet::InitStatsForLevel() failed for creature (Entry: %u)!", creatureTarget->GetEntry());
+        delete pet;
+        return;
+    }
+
+    pet->GetCharmInfo()->SetPetNumber(sObjectMgr.GeneratePetNumber(), true);
+    pet->SetHealth(pet->GetMaxHealth());
+
+    // "kill" original creature
+    creatureTarget->ForcedDespawn();
+
+    // prepare visual effect for levelup
+    pet->SetUInt32Value(UNIT_FIELD_LEVEL, creatureTarget->getLevel() - 1);
+
+    // add to world
+    pet->GetMap()->Add((Creature*)pet);
+    pet->AIM_Initialize();
+    pet->InitPetCreateSpells();
+
+    // visual effect for levelup
+    pet->SetUInt32Value(UNIT_FIELD_LEVEL, creatureTarget->getLevel());
+
+    // caster have pet now
+    plr->SetPet(pet);
+
+    pet->SavePetToDB(PET_SAVE_AS_CURRENT);
+    plr->PetSpellInitialize();
+}
+
+/**
+ * @brief Summons, recalls, or replaces the caster's pet.
+ *
+ * @param eff_idx The summon effect index.
+ */
+void Spell::EffectSummonPet(SpellEffectIndex eff_idx)
+{
+    uint32 petentry = m_spellInfo->EffectMiscValue[eff_idx];
+
+    Pet* OldSummon = m_caster->GetPet();
+
+    // if pet requested type already exist
+    if (OldSummon)
+    {
+        if ((petentry == 0 || OldSummon->GetEntry() == petentry) && OldSummon->getPetType() != SUMMON_PET)
+        {
+            // pet in corpse state can't be summoned
+            if (OldSummon->IsDead())
+            {
+                return;
+            }
+
+            OldSummon->GetMap()->Remove((Creature*)OldSummon, false);
+
+            float px, py, pz;
+            m_caster->GetClosePoint(px, py, pz, OldSummon->GetObjectBoundingRadius());
+
+            OldSummon->Relocate(px, py, pz, OldSummon->GetOrientation());
+            m_caster->GetMap()->Add((Creature*)OldSummon);
+
+            if (m_caster->GetTypeId() == TYPEID_PLAYER && OldSummon->isControlled())
+            {
+                ((Player*)m_caster)->PetSpellInitialize();
+            }
+            return;
+        }
+
+        if (m_caster->GetTypeId() == TYPEID_PLAYER)
+        {
+            OldSummon->Unsummon(OldSummon->getPetType() == HUNTER_PET ? PET_SAVE_AS_DELETED : PET_SAVE_NOT_IN_SLOT, m_caster);
+        }
+        else
+        {
+            return;
+        }
+    }
+
+    CreatureInfo const* cInfo = petentry ? ObjectMgr::GetCreatureTemplate(petentry) : NULL;
+
+    // == 0 in case call current pet, check only real summon case
+    if (petentry && !cInfo)
+    {
+        sLog.outErrorDb("EffectSummonPet: creature entry %u not found for spell %u.", petentry, m_spellInfo->Id);
+        return;
+    }
+
+    Pet* NewSummon = new Pet;
+
+    // petentry==0 for hunter "call pet" (current pet summoned if any)
+    if (m_caster->GetTypeId() == TYPEID_PLAYER && NewSummon->LoadPetFromDB((Player*)m_caster, petentry))
+    {
+        if (NewSummon->getPetType() == SUMMON_PET)
+        {
+            // Remove Demonic Sacrifice auras (known pet)
+            Unit::AuraList const& auraClassScripts = m_caster->GetAurasByType(SPELL_AURA_OVERRIDE_CLASS_SCRIPTS);
+            for (Unit::AuraList::const_iterator itr = auraClassScripts.begin(); itr != auraClassScripts.end();)
+            {
+                if ((*itr)->GetModifier()->m_miscvalue == 2228)
+                {
+                    m_caster->RemoveAurasDueToSpell((*itr)->GetId());
+                    itr = auraClassScripts.begin();
+                }
+                else
+                {
+                    ++itr;
+                }
+            }
+        }
+
+        return;
+    }
+
+    // not error in case fail hunter call pet
+    if (!petentry)
+    {
+        delete NewSummon;
+        return;
+    }
+
+    CreatureCreatePos pos(m_caster, m_caster->GetOrientation());
+
+    Map* map = m_caster->GetMap();
+    uint32 pet_number = sObjectMgr.GeneratePetNumber();
+    if (!NewSummon->Create(map->GenerateLocalLowGuid(HIGHGUID_PET), pos, cInfo, pet_number))
+    {
+        delete NewSummon;
+        return;
+    }
+
+    NewSummon->SetRespawnCoord(pos);
+
+    uint32 petlevel = std::max(m_caster->getLevel() + m_spellInfo->EffectMultipleValue[eff_idx], 1.0f);
+    NewSummon->setPetType(SUMMON_PET);
+
+    uint32 faction = m_caster->getFaction();
+    if (m_caster->GetTypeId() == TYPEID_UNIT)
+    {
+        if (((Creature*)m_caster)->IsTotem())
+        {
+            NewSummon->GetCharmInfo()->SetReactState(REACT_AGGRESSIVE);
+        }
+        else
+        {
+            NewSummon->GetCharmInfo()->SetReactState(REACT_DEFENSIVE);
+        }
+    }
+
+    NewSummon->SetOwnerGuid(m_caster->GetObjectGuid());
+    NewSummon->SetCreatorGuid(m_caster->GetObjectGuid());
+    NewSummon->SetUInt32Value(UNIT_NPC_FLAGS, UNIT_NPC_FLAG_NONE);
+    NewSummon->setFaction(faction);
+    NewSummon->SetUInt32Value(UNIT_FIELD_PET_NAME_TIMESTAMP, uint32(time(NULL)));
+    NewSummon->SetUInt32Value(UNIT_FIELD_PETEXPERIENCE, 0);
+    NewSummon->SetUInt32Value(UNIT_FIELD_PETNEXTLEVELEXP, 1000);
+    NewSummon->SetUInt32Value(UNIT_CREATED_BY_SPELL, m_spellInfo->Id);
+
+    NewSummon->GetCharmInfo()->SetPetNumber(pet_number, true);
+    // this enables pet details window (Shift+P)
+
+    if (m_caster->IsPvP())
+    {
+        NewSummon->SetPvP(true);
+    }
+
+    NewSummon->InitStatsForLevel(petlevel, m_caster);
+    NewSummon->InitPetCreateSpells();
+
+    if (NewSummon->getPetType() == SUMMON_PET)
+    {
+        // Remove Demonic Sacrifice auras (new pet)
+        Unit::AuraList const& auraClassScripts = m_caster->GetAurasByType(SPELL_AURA_OVERRIDE_CLASS_SCRIPTS);
+        for (Unit::AuraList::const_iterator itr = auraClassScripts.begin(); itr != auraClassScripts.end();)
+        {
+            if ((*itr)->GetModifier()->m_miscvalue == 2228)
+            {
+                m_caster->RemoveAurasDueToSpell((*itr)->GetId());
+                itr = auraClassScripts.begin();
+            }
+            else
+            {
+                ++itr;
+            }
+        }
+    }
+
+    if (m_caster->GetTypeId() == TYPEID_PLAYER && NewSummon->getPetType() == SUMMON_PET)
+    {
+        // generate new name for summon pet
+        std::string new_name = sObjectMgr.GeneratePetName(petentry);
+        if (!new_name.empty())
+        {
+            NewSummon->SetName(new_name);
+        }
+    }
+
+    if (NewSummon->getPetType() == HUNTER_PET)
+    {
+        NewSummon->RemoveByteFlag(UNIT_FIELD_BYTES_2, 2, UNIT_CAN_BE_RENAMED);
+        NewSummon->SetByteFlag(UNIT_FIELD_BYTES_2, 2, UNIT_CAN_BE_ABANDONED);
+    }
+
+    NewSummon->SetHealth(NewSummon->GetMaxHealth());
+    NewSummon->SetPower(POWER_MANA, NewSummon->GetMaxPower(POWER_MANA));
+
+    map->Add((Creature*)NewSummon);
+    NewSummon->AIM_Initialize();
+
+    m_caster->SetPet(NewSummon);
+    DEBUG_LOG("New Pet has guid %u", NewSummon->GetGUIDLow());
+
+    if (m_caster->GetTypeId() == TYPEID_PLAYER)
+    {
+        NewSummon->SavePetToDB(PET_SAVE_AS_CURRENT);
+        ((Player*)m_caster)->PetSpellInitialize();
+    }
+}
+
+/**
+ * @brief Teaches a new spell to the caster's pet.
+ *
+ * @param eff_idx The effect index containing the learned spell id.
+ */
+void Spell::EffectLearnPetSpell(SpellEffectIndex eff_idx)
+{
+    if (m_caster->GetTypeId() != TYPEID_PLAYER)
+    {
+        return;
+    }
+
+    Player* _player = (Player*)m_caster;
+
+    Pet* pet = _player->GetPet();
+    if (!pet)
+    {
+        return;
+    }
+    if (!pet->IsAlive())
+    {
+        return;
+    }
+
+    SpellEntry const* learn_spellproto = sSpellStore.LookupEntry(m_spellInfo->EffectTriggerSpell[eff_idx]);
+    if (!learn_spellproto)
+    {
+        return;
+    }
+
+    pet->SetTP(pet->m_TrainingPoints - pet->GetTPForSpell(learn_spellproto->Id));
+    pet->learnSpell(learn_spellproto->Id);
+
+    pet->SavePetToDB(PET_SAVE_AS_CURRENT);
+    _player->PetSpellInitialize();
+
+    if (WorldObject const* caster = GetCastingObject())
+    {
+        DEBUG_LOG("Spell: %s has learned spell %u from %s", pet->GetGuidStr().c_str(), learn_spellproto->Id, caster->GetGuidStr().c_str());
+    }
+}
+
+/**
+ * @brief Prepares taunt threat adjustments before the taunt aura is applied.
+ */
+void Spell::EffectTaunt(SpellEffectIndex /*eff_idx*/)
+{
+    if (!unitTarget)
+    {
+        return;
+    }
+
+    // this effect use before aura Taunt apply for prevent taunt already attacking target
+    // for spell as marked "non effective at already attacking target"
+    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
+    {
+        if (unitTarget->getVictim() == m_caster)
+        {
+            SendCastResult(SPELL_FAILED_DONT_REPORT);
+            return;
+        }
+    }
+
+    // Also use this effect to set the taunter's threat to the taunted creature's highest value
+    if (unitTarget->CanHaveThreatList() && unitTarget->GetThreatManager().getCurrentVictim())
+    {
+        unitTarget->GetThreatManager().addThreat(m_caster, unitTarget->GetThreatManager().getCurrentVictim()->getThreat());
+    }
+}
+
+/**
+ * @brief Computes weapon-based spell damage and accumulates it for application.
+ *
+ * @param eff_idx The weapon damage effect index.
+ */
+void Spell::EffectWeaponDmg(SpellEffectIndex eff_idx)
+{
+    if (!unitTarget)
+    {
+        return;
+    }
+    if (!unitTarget->IsAlive())
+    {
+        return;
+    }
+
+    // multiple weapon dmg effect workaround
+    // execute only the last weapon damage
+    // and handle all effects at once
+    for (int j = 0; j < MAX_EFFECT_INDEX; ++j)
+    {
+        switch (m_spellInfo->Effect[j])
+        {
+            case SPELL_EFFECT_WEAPON_DAMAGE:
+            case SPELL_EFFECT_WEAPON_DAMAGE_NOSCHOOL:
+            case SPELL_EFFECT_NORMALIZED_WEAPON_DMG:
+            case SPELL_EFFECT_WEAPON_PERCENT_DAMAGE:
+                if (j < int(eff_idx))                       // we must calculate only at last weapon effect
+                {
+                    return;
+                }
+                break;
+        }
+    }
+
+    // some spell specific modifiers
+    bool customBonusDamagePercentMod = false;
+    float bonusDamagePercentMod  = 1.0f;                    // applied to fixed effect damage bonus if set customBonusDamagePercentMod
+    float weaponDamagePercentMod = 1.0f;                    // applied to weapon damage (and to fixed effect damage bonus if customBonusDamagePercentMod not set
+    float totalDamagePercentMod  = 1.0f;                    // applied to final bonus+weapon damage
+    bool normalized = false;
+
+    int32 spell_bonus = 0;                                  // bonus specific for spell
+
+    switch (m_spellInfo->SpellFamilyName)
+    {
+        case SPELLFAMILY_WARRIOR:
+        {
+            // Devastate
+            if (m_spellInfo->SpellVisual == 671 && m_spellInfo->SpellIconID == 1508)
+            {
+                customBonusDamagePercentMod = true;
+                bonusDamagePercentMod = 0.0f;               // only applied if auras found
+
+                // Sunder Armor
+                Aura* sunder = unitTarget->GetAura(SPELL_AURA_MOD_RESISTANCE, SPELLFAMILY_WARRIOR, UI64LIT(0x0000000000004000), m_caster->GetObjectGuid());
+
+                // Devastate bonus and sunder armor refresh, additional threat
+                if (sunder)
+                {
+                    sunder->GetHolder()->RefreshHolder();
+
+                    // 100% * stack
+                    bonusDamagePercentMod += 1.0f * sunder->GetStackAmount();
+
+                    // 25 * stack
+                    unitTarget->AddThreat(m_caster, 25.0f * sunder->GetStackAmount(), false, GetSpellSchoolMask(m_spellInfo), m_spellInfo);
+                }
+            }
+            break;
+        }
+        case SPELLFAMILY_ROGUE:
+        {
+            // Ambush
+            if (m_spellInfo->SpellFamilyFlags & UI64LIT(0x00000200))
+            {
+                customBonusDamagePercentMod = true;
+                bonusDamagePercentMod = 2.5f;               // 250%
+            }
+            // Mutilate (for each hand)
+            else if (m_spellInfo->SpellFamilyFlags & UI64LIT(0x600000000))
+            {
+                bool found = false;
+                // fast check
+                if (unitTarget->HasAuraState(AURA_STATE_DEADLY_POISON))
+                {
+                    found = true;
+                }
+                // full aura scan
+                else
+                {
+                    Unit::SpellAuraHolderMap const& auras = unitTarget->GetSpellAuraHolderMap();
+                    for (Unit::SpellAuraHolderMap::const_iterator itr = auras.begin(); itr != auras.end(); ++itr)
+                    {
+                        if (itr->second->GetSpellProto()->Dispel == DISPEL_POISON)
+                        {
+                            found = true;
+                            break;
+                        }
+                    }
+                }
+
+                if (found)
+                {
+                    totalDamagePercentMod *= 1.5f;          // 150% if poisoned
+                }
+            }
+            break;
+        }
+        case SPELLFAMILY_PALADIN:
+        {
+            // Seal of Command - receive benefit from Spell Damage and Healing
+            if (m_spellInfo->SpellFamilyFlags & UI64LIT(0x00000002000000))
+            {
+                spell_bonus += int32(0.20f * m_caster->SpellBaseDamageBonusDone(GetSpellSchoolMask(m_spellInfo)));
+                spell_bonus += int32(0.29f * unitTarget->SpellBaseDamageBonusTaken(GetSpellSchoolMask(m_spellInfo)));
+            }
+            break;
+        }
+        case SPELLFAMILY_SHAMAN:
+        {
+            // Skyshatter Harness item set bonus
+            // Stormstrike
+            if (m_spellInfo->SpellFamilyFlags & UI64LIT(0x001000000000))
+            {
+                Unit::AuraList const& m_OverrideClassScript = m_caster->GetAurasByType(SPELL_AURA_OVERRIDE_CLASS_SCRIPTS);
+                for (Unit::AuraList::const_iterator citr = m_OverrideClassScript.begin(); citr != m_OverrideClassScript.end(); ++citr)
+                {
+                    // Stormstrike AP Buff
+                    if ((*citr)->GetModifier()->m_miscvalue == 5634)
+                    {
+                        m_caster->CastSpell(m_caster, 38430, true, NULL, *citr);
+                        break;
+                    }
+                }
+            }
+            break;
+        }
+    }
+
+    int32 fixed_bonus = 0;
+    for (int j = 0; j < MAX_EFFECT_INDEX; ++j)
+    {
+        switch (m_spellInfo->Effect[j])
+        {
+            case SPELL_EFFECT_WEAPON_DAMAGE:
+            case SPELL_EFFECT_WEAPON_DAMAGE_NOSCHOOL:
+                fixed_bonus += CalculateDamage(SpellEffectIndex(j), unitTarget);
+                break;
+            case SPELL_EFFECT_NORMALIZED_WEAPON_DMG:
+                fixed_bonus += CalculateDamage(SpellEffectIndex(j), unitTarget);
+                normalized = true;
+                break;
+            case SPELL_EFFECT_WEAPON_PERCENT_DAMAGE:
+                weaponDamagePercentMod *= float(CalculateDamage(SpellEffectIndex(j), unitTarget)) / 100.0f;
+
+                // applied only to prev.effects fixed damage
+                if (customBonusDamagePercentMod)
+                {
+                    fixed_bonus = int32(fixed_bonus * bonusDamagePercentMod);
+                }
+                else
+                {
+                    fixed_bonus = int32(fixed_bonus * weaponDamagePercentMod);
+                }
+                break;
+            default:
+                break;                                      // not weapon damage effect, just skip
+        }
+    }
+
+    // non-weapon damage
+    int32 bonus = spell_bonus + fixed_bonus;
+
+    // apply to non-weapon bonus weapon total pct effect, weapon total flat effect included in weapon damage
+    if (bonus)
+    {
+        UnitMods unitMod;
+        switch (m_attackType)
+        {
+            default:
+            case BASE_ATTACK:   unitMod = UNIT_MOD_DAMAGE_MAINHAND; break;
+            case OFF_ATTACK:    unitMod = UNIT_MOD_DAMAGE_OFFHAND;  break;
+            case RANGED_ATTACK: unitMod = UNIT_MOD_DAMAGE_RANGED;   break;
+        }
+
+        float weapon_total_pct  = m_caster->GetModifierValue(unitMod, TOTAL_PCT);
+        bonus = int32(bonus * weapon_total_pct);
+    }
+
+    // + weapon damage with applied weapon% dmg to base weapon damage in call
+    bonus += int32(m_caster->CalculateDamage(m_attackType, normalized) * weaponDamagePercentMod);
+
+    // total damage
+    bonus = int32(bonus * totalDamagePercentMod);
+
+    // prevent negative damage
+    m_damage += uint32(bonus > 0 ? bonus : 0);
+
+    // Hemorrhage
+    if (m_spellInfo->IsFitToFamily(SPELLFAMILY_ROGUE, UI64LIT(0x0000000002000000)))
+    {
+        if (m_caster->GetTypeId() == TYPEID_PLAYER)
+        {
+            ((Player*)m_caster)->AddComboPoints(unitTarget, 1);
+        }
+    }
+    // Mangle (Cat): CP
+    else if (m_spellInfo->IsFitToFamily(SPELLFAMILY_DRUID, UI64LIT(0x0000040000000000)))
+    {
+        if (m_caster->GetTypeId() == TYPEID_PLAYER)
+        {
+            ((Player*)m_caster)->AddComboPoints(unitTarget, 1);
+        }
+    }
+}
